feat(ST-46): components refactor to enhance usability
