+++ modules/classroom/containers/classroomContainer/ClassroomContainer.tsx
@@ -5,6 +5,7 @@ import { useRouter } from "next/router";
 import { Tabs } from "@mantine/core";
 
 import PeopleContainer from "@/modules/classroom/containers/peopleContainer";
+import StreamContainer from "@/modules/classroom/containers/streamContainer";
 import LoadingComponent from "@/shared/components/LoadingComponent";
 import { useLazyGetOneClassroomQuery } from "@/shared/redux/rtk-apis/classrooms/classrooms.api";
 
@@ -36,7 +37,7 @@ export default function ClassroomContainer() {
     console.log(classroomData.participants?.length);
     return (
       <div className="mx-8">
-        <Tabs defaultValue="classroom">
+        <Tabs defaultValue="stream">
           <Tabs.List grow className="text-white justify-center">
             <Tabs.Tab value="stream" color="teal" className="hover:bg-transparent">
               <span className="text-white">Stream</span>
@@ -49,7 +50,9 @@ export default function ClassroomContainer() {
             </Tabs.Tab>
           </Tabs.List>
 
-          <Tabs.Panel value="stream">stream tab content</Tabs.Panel>
+          <Tabs.Panel value="stream">
+            <StreamContainer />
+          </Tabs.Panel>
           <Tabs.Panel value="classwork">classwork tab content</Tabs.Panel>
           <Tabs.Panel value="people">
             <PeopleContainer
